requirements.txt
flask
flask_sqlalchemy
flask_restful
flask_bcrypt
flask_jwt_extended
pymysql

